Time Error
Using to store the milliseconds data eventually saturated the integer
and the lights began blinking constantly. Changing to long-type solved
this problem.

+++ Lab_2_Exercise_2/Lab_2_Exercise_2.ino
@@ -36,7 +36,7 @@ void loop() {
   const int pot = A0;
   
   //Declares some initial variables
-  int steadyTime = millis();
+  long steadyTime = millis();
   int state = 1;
   int timeDifference = 500;
   int ledDuration = 3;
